Support for multiple embeds. Initial commit
MAY NOT WORK
May contain some unrelated changes

+++ inc/func/apc.php
@@ -11,12 +11,14 @@ function clearBlotterCache() {
  *
  * @param integer $id Post ID
  * @param string $board Board name
+ * @param boolean $skipdb Whether or not deleting from reports table should be skipped
  */
-function clearPostCache($id, $board) {
+function clearPostCache($id, $board, $skipdb=false) {
 	global $tc_db;
 	if (KU_APC) {
 		apc_delete('post|' . $board . '|' . $id);
 	}
-	$tc_db->Execute("DELETE FROM `" . KU_DBPREFIX . "reports` WHERE `postid` = " . $id . " AND `board` = " . $tc_db->qstr($board));
+  if (!$skipdb)
+	 $tc_db->Execute("DELETE FROM `" . KU_DBPREFIX . "reports` WHERE `postid` = " . $id . " AND `board` = " . $tc_db->qstr($board));
 }
 ?>
